refactor(backend): close direct prompt compat exports
